test(outreach): fix the two funnel-scan tests this PR broke, one harness
The three empty outcomes now log at DEBUG, but `TestOutreachFunnelSourcing`
in test_network_activation.py already asserted two of those exact messages
via `caplog` — and the `cqc-lem` logger sits at INFO, so both assertions saw
an empty capture and Unit Tests went red. They now raise the capture level
for that logger, which is the behaviour under test.

The new escalation-contract tests moved here too. They had been added to
test_outreach_funnel.py behind a second copy of the `_scan_funnel` mock
harness, aimed at the same task the harness in this file already drives —
so the file that actually covers `scan_outreach_funnel_targets` was never
run. One harness, extended with `profile_exc` so the roster-sourcing
failure path can assert it still warns.
